+++ CHANGELOG.md
@@ -53,6 +53,8 @@ with the `YY.MINOR.MICRO` scheme.
 
 - DdlAbort raft entry now contains the error information.
 
+- Add `promotion_vclock` column to `_pico_replicaset` table.
+
 ### Lua API
 
 - Update `pico.LUA_API_VERSION`: `3.1.0` -> `4.0.0`